+++ lib/src/widget/component.dart
@@ -25,38 +25,40 @@ class Component extends StatelessWidget {
       top: position.dy,
       width: canvasState.canvasFinalScale() * componentData.size.width,
       height: canvasState.canvasFinalScale() * componentData.size.height,
-      child: Stack(
-        clipBehavior: Clip.none,
-        children: [
-          Positioned(
-            left: 0,
-            top: 0,
-            width: componentData.size.width * canvasState.canvasFinalScale(),
-            height: componentData.size.height * canvasState.canvasFinalScale(),
-            child: Listener(
-              onPointerSignal: (PointerSignalEvent event) {
-                policy.onComponentPointerSignal(componentData.id, event);
-              },
-              child: GestureDetector(
-                //behavior: HitTestBehavior.translucent,
-                child: policy.showComponentBody(componentData) ?? Container(),
-                onTap: () => policy.onComponentTap(componentData.id),
-                onTapDown: (details) => policy.onComponentTapDown(componentData.id, details),
-                onTapUp: (details) => policy.onComponentTapUp(componentData.id, details),
-                onTapCancel: () => policy.onComponentTapCancel(componentData.id),
-                onScaleStart: (details) => policy.onComponentScaleStart(componentData.id, details),
-                onScaleUpdate: (details) => policy.onComponentScaleUpdate(componentData.id, details),
-                onScaleEnd: (details) => policy.onComponentScaleEnd(componentData.id, details),
-                onLongPress: () => policy.onComponentLongPress(componentData.id),
-                onLongPressStart: (details) => policy.onComponentLongPressStart(componentData.id, details),
-                onLongPressMoveUpdate: (details) => policy.onComponentLongPressMoveUpdate(componentData.id, details),
-                onLongPressEnd: (details) => policy.onComponentLongPressEnd(componentData.id, details),
-                onLongPressUp: () => policy.onComponentLongPressUp(componentData.id),
+      child: RepaintBoundary(
+        child: Stack(
+          clipBehavior: Clip.none,
+          children: [
+            Positioned(
+              left: 0,
+              top: 0,
+              width: componentData.size.width * canvasState.canvasFinalScale(),
+              height: componentData.size.height * canvasState.canvasFinalScale(),
+              child: Listener(
+                onPointerSignal: (PointerSignalEvent event) {
+                  policy.onComponentPointerSignal(componentData.id, event);
+                },
+                child: GestureDetector(
+                  //behavior: HitTestBehavior.translucent,
+                  child: policy.showComponentBody(componentData) ?? Container(),
+                  onTap: () => policy.onComponentTap(componentData.id),
+                  onTapDown: (details) => policy.onComponentTapDown(componentData.id, details),
+                  onTapUp: (details) => policy.onComponentTapUp(componentData.id, details),
+                  onTapCancel: () => policy.onComponentTapCancel(componentData.id),
+                  onScaleStart: (details) => policy.onComponentScaleStart(componentData.id, details),
+                  onScaleUpdate: (details) => policy.onComponentScaleUpdate(componentData.id, details),
+                  onScaleEnd: (details) => policy.onComponentScaleEnd(componentData.id, details),
+                  onLongPress: () => policy.onComponentLongPress(componentData.id),
+                  onLongPressStart: (details) => policy.onComponentLongPressStart(componentData.id, details),
+                  onLongPressMoveUpdate: (details) => policy.onComponentLongPressMoveUpdate(componentData.id, details),
+                  onLongPressEnd: (details) => policy.onComponentLongPressEnd(componentData.id, details),
+                  onLongPressUp: () => policy.onComponentLongPressUp(componentData.id),
+                ),
               ),
             ),
-          ),
-          policy.showCustomWidgetWithComponentData(context, componentData),
-        ],
+            policy.showCustomWidgetWithComponentData(context, componentData),
+          ],
+        ),
       ),
     );
   }